+++ CHANGELOG.md
@@ -10,7 +10,9 @@
 * Added `operation` parameter to the following annotations: `@Adjacency`, `@Incidence`, `@Property`. Setting the
   parameter will override the auto discovery of the method prefix previously used to discovery the operation of the
   method.
-* Methods annotated with the `Adjacency` annotation can now return a `List` or a `Set` in addition to the usual
+* Methods annotated with the `@Adjacency` annotation can now return a `List` or a `Set` in addition to the usual
+  `Iterator` return type.
+* Methods annotated with the `@Incidence` annotation can now return a `List` or a `Set` in addition to the usual
   `Iterator` return type.
 
 ## 3.1.0